requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.1
pydantic>=2.5.2
